[NFC] Extract Pipeline into support library
This extracts the pipeline description into a new Support library. This
is preprataion for generating an Image library to work with images more
formally, which will in turn depend on the Pipeline data structures.
