+++ lib/API/DX/Device.cpp
@@ -24,7 +24,7 @@
 #include "DXFeatures.h"
 #include "API/Capabilities.h"
 #include "API/Device.h"
-#include "API/Pipeline.h"
+#include "Support/Pipeline.h"
 #include "WinError.h"
 
 #include "llvm/ADT/SmallVector.h"
