+++ lib/CMakeLists.txt
@@ -1,1 +1,2 @@
 add_subdirectory(API)
+add_subdirectory(Support)
