+++ lib/Support/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_hlsl_library(Support Pipeline.cpp)
