+++ tools/offloader/CMakeLists.txt
@@ -4,4 +4,4 @@ add_hlsl_tool(offloader
 
 target_include_directories(offloader PRIVATE SYSTEM BEFORE "${HLSLTEST_BINARY_DIR}/third-party/libpng/")
 
-target_link_libraries(offloader PRIVATE LLVMSupport HLSLTestAPI png_static)
+target_link_libraries(offloader PRIVATE LLVMSupport HLSLTestAPI HLSLTestSupport png_static)
